+++ pygametemplate/console.py
@@ -4,8 +4,15 @@ from __future__ import absolute_import
 from pygametemplate.hotkey import Hotkey
 
 class Console(object):
-    def __init__(self, game):
+
+    def __init__(self, game, toggle_fps_hotkey=None):
+        """Create a Console instance.
+
+        `toggle_fps_hotkey` defaults to Ctrl+F.
+        """
         self.game = game
+
+        toggle_fps_hotkey = toggle_fps_hotkey or Hotkey(self.game, "f", ctrl=True)
         try:
             self.font = self.game.pygame.font.SysFont("consolas", 15)
             self.text_colour = (255, 255, 255)  # white
@@ -14,7 +21,7 @@ class Console(object):
             self.fps_coordinates = (game.width - self.font.size("FPS: 000")[0], 0)
 
             self.hotkeys = {    # (hotkey condition, function)
-                "toggle fps": (Hotkey(self.game, "f", ctrl=True).pressed, self.toggle_fps)
+                "toggle fps": (toggle_fps_hotkey.pressed, self.toggle_fps)
             }
         except Exception:
             self.game.log("Failed to initialise console object")